requirements.txt
python-telegram-bot[webhooks]==21.10
flask==3.1.2
supabase==2.27.0
python-dotenv==1.1.0
fpdf2==2.8.5
qrcode==8.2
Pillow==12.0.0
aiohttp==3.13.2
pytz==2025.2
yt-dlp==2025.12.8
openai==2.14.0
shazamio==0.8.1
pydantic==2.12.3
pydantic-settings==2.10.1
redis==7.3.0
